attendance-system/attendance_backend/requirements.txt
# requirements.txt - VERSIÓN SIN PROBLEMAS
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.0.1
# cmake==3.27.7  # No necesario si no instalamos dlib manualmente
# dlib==19.24.2  # Se instala automáticamente con face-recognition